Refactor AppDelegate launch logic
